+++ src/client/src/services/hud/appearance.outlet.ts
@@ -1,0 +1,5 @@
+export class AppearanceOutlet {
+
+    constructor() {}
+
+}