+++ src/client/src/services/picker/element-holder.ts
@@ -1,4 +1,12 @@
+import { ClientEvent, connectEvent } from "../../entities/client-event";
+
+const enum ElementHolderEvents {
+    setElements = 'setElements',
+}
+
 export class ElementHolder {
+
+    public readonly [ElementHolderEvents.setElements] = new ClientEvent<SVGElement[]>();
 
     private box = Array<SVGElement>();
 
@@ -11,6 +19,12 @@ export class ElementHolder {
     set elements(val: SVGElement[]) {
         this.box = val;
         this.callbacks.forEach(callback => callback(val));
+        this.fireElements(val);
+    }
+
+    @connectEvent(ElementHolderEvents.setElements)
+    fireElements(elements: SVGElement[]): SVGElement[] {
+        return elements;
     }
 
     addListener(callback: (element: SVGElement[]) => void) {