+++ src/shared/pipes/artboard.pipe.ts
@@ -6,3 +6,5 @@ import { Pipe } from "../services/pipe/pipe";
 export type ArtboardRequest = {property: string; value?: string};
 export type ArtboardResponse = {value: string | null};
 export const artboardPipe = new Pipe<ArtboardRequest, ArtboardResponse, 'artboard'>('artboard');
+
+export const artboardInversePipe = new Pipe<ArtboardRequest, ArtboardResponse, 'artboard-inverse'>('artboard-inverse');